requirements.txt
autopep8==1.6.0
Django==2.2.10
django-nested-admin==3.4.0
djangorestframework==3.13.1
pycodestyle==2.8.0
python-monkey-business==1.0.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
